Separate manageClient into more files for readability
