+++ Kentico.Kontent.Management/ManagementClient.ProjectUser.cs
@@ -1,0 +1,34 @@
+﻿using Kentico.Kontent.Management.Models.Shared;
+using Kentico.Kontent.Management.Models.Users;
+using System;
+using System.Net.Http;
+using System.Threading.Tasks;
+
+namespace Kentico.Kontent.Management;
+
+public partial class ManagementClient : IManagementClient
+{
+    /// <inheritdoc />
+    public async Task<UserModel> InviteUserIntoProjectAsync(UserInviteModel invitation)
+    {
+        if (invitation == null)
+        {
+            throw new ArgumentNullException(nameof(invitation));
+        }
+
+        var endpointUrl = _urlBuilder.BuildUsersUrl();
+        return await _actionInvoker.InvokeMethodAsync<UserInviteModel, UserModel>(endpointUrl, HttpMethod.Post, invitation);
+    }
+
+    /// <inheritdoc />
+    public async Task<UserModel> ModifyUsersRolesAsync(UserIdentifier identifier, UserModel user)
+    {
+        if (identifier == null)
+        {
+            throw new ArgumentNullException(nameof(identifier));
+        }
+
+        var endpointUrl = _urlBuilder.BuildModifyUsersRoleUrl(identifier);
+        return await _actionInvoker.InvokeMethodAsync<UserModel, UserModel>(endpointUrl, HttpMethod.Put, user);
+    }
+}
