+++ Kentico.Kontent.Management/ManagementClient.Workflow.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Net.Http;
 using System.Threading.Tasks;
+using Kentico.Kontent.Management.Models.LanguageVariants;
 using Kentico.Kontent.Management.Models.Shared;
 using Kentico.Kontent.Management.Models.Workflow;
 
@@ -29,5 +30,114 @@ public sealed partial class ManagementClient
 
         var endpointUrl = _urlBuilder.BuildWorkflowsUrl(identifier);
         await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Delete);
+    }
+
+    /// <inheritdoc />
+    public async Task ChangeLanguageVariantWorkflowAsync(LanguageVariantIdentifier identifier, WorkflowStepIdentifier workflowStepIdentifier)
+    {
+        if (identifier == null)
+        {
+            throw new ArgumentNullException(nameof(identifier));
+        }
+
+        if (workflowStepIdentifier == null)
+        {
+            throw new ArgumentNullException(nameof(workflowStepIdentifier));
+        }
+
+        var endpointUrl = _urlBuilder.BuildWorkflowChangeUrl(identifier);
+
+        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put, workflowStepIdentifier);
+    }
+
+    /// <inheritdoc />
+    public async Task PublishLanguageVariantAsync(LanguageVariantIdentifier identifier)
+    {
+        if (identifier == null)
+        {
+            throw new ArgumentNullException(nameof(identifier));
+        }
+
+        var endpointUrl = _urlBuilder.BuildPublishVariantUrl(identifier);
+
+        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put);
+    }
+
+    /// <inheritdoc />
+    public async Task SchedulePublishingOfLanguageVariantAsync(LanguageVariantIdentifier identifier, ScheduleModel scheduleModel)
+    {
+        if (identifier == null)
+        {
+            throw new ArgumentNullException(nameof(identifier));
+        }
+
+        var endpointUrl = _urlBuilder.BuildPublishVariantUrl(identifier);
+
+        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put, scheduleModel);
+    }
+
+    /// <inheritdoc />
+    public async Task CancelPublishingOfLanguageVariantAsync(LanguageVariantIdentifier identifier)
+    {
+        if (identifier == null)
+        {
+            throw new ArgumentNullException(nameof(identifier));
+        }
+
+        var endpointUrl = _urlBuilder.BuildCancelPublishingVariantUrl(identifier);
+
+        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put);
+    }
+
+    /// <inheritdoc />
+    public async Task UnpublishLanguageVariantAsync(LanguageVariantIdentifier identifier)
+    {
+        if (identifier == null)
+        {
+            throw new ArgumentNullException(nameof(identifier));
+        }
+
+        var endpointUrl = _urlBuilder.BuildUnpublishVariantUrl(identifier);
+
+        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put);
+    }
+
+    /// <inheritdoc />
+    public async Task CancelUnpublishingOfLanguageVariantAsync(LanguageVariantIdentifier identifier)
+    {
+        if (identifier == null)
+        {
+            throw new ArgumentNullException(nameof(identifier));
+        }
+
+        var endpointUrl = _urlBuilder.BuildCancelUnpublishingVariantUrl(identifier);
+
+        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put);
+    }
+
+    /// <inheritdoc />
+    public async Task ScheduleUnpublishingOfLanguageVariantAsync(LanguageVariantIdentifier identifier, ScheduleModel scheduleModel)
+    {
+        if (identifier == null)
+        {
+            throw new ArgumentNullException(nameof(identifier));
+        }
+
+        var endpointUrl = _urlBuilder.BuildUnpublishVariantUrl(identifier);
+
+        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put, scheduleModel);
+    }
+
+    /// <inheritdoc />
+    public async Task CreateNewVersionOfLanguageVariantAsync(LanguageVariantIdentifier identifier)
+    {
+        if (identifier == null)
+        {
+            throw new ArgumentNullException(nameof(identifier));
+        }
+
+        var endpointUrl = _urlBuilder.BuildNewVersionVariantUrl(identifier);
+
+        await _actionInvoker.InvokeMethodAsync(endpointUrl, HttpMethod.Put);
     }
 }
